HSTTWO-3836 Support _index_  siteMapItem matching
I've added to the unit test content setup new sitemap structure containing _index_ sitemap items. Next
to the extra sitemap items,  I added content that is referenced by the _index_ sitemap items.

The matching now checks whether below the matched sitemap item a _index_ sitemap item exists. If
the _index_ item exists *and* it points to an existing document, we return the _index_ sitemap item
as resolved sitemap item. Note that the check whether the document exists happens with the jcr session
that is available during the matching phase via the HstRequestContext. This is not 100% guaranteed the
same jcr session that is later used during the processing phase. Potentially, the processing phase returns
a jcr session via the HstRequestContext that is a different one then during matching and thus there is a
theoretical chance that the document beloning to the _index_ document cannot be read by that session.
However, this is 99.99% a theoretical situation which is extremely unlikely in a real application. Apart
from that, there is no way to solve this otherwise within the current architecture.