CMS-10518 Don't redraw an editor tab when activated
While redrawing ensures the latest backend state is shown, it
'resets' the frontend state. This may have unforeseen consequences
for many field types, so better keep the old behavior. It's
already a fishy sign that the backend state changed while a document
is being edited since the CMS UI prevents that from happening.