HSTTWO-3875 keep track of uncacheable keys
for example when some url results in that an HstComponent sets a cookie, which
takes place *after* the page caching valve, the response is uncacheable. In that case, not a
second level lookup has to be done for such keys. Hence, we keep track of the last 1000 (lru map)
uncacheable keys for performance reasons