CHANNELMGR-898 Reload document type when save fails on invalid fields
When the backend fails to save because of invalid fields, it returns
the document with the fields-in-error marked. The only real scenario
in which this can currently happen is a non-required string field
that, while the editor was open, was made required by another user
that changes the document type.

In that case we reload the document type. Since validation for
required fields is also client-side, Angular will then render the
suddenly invalid fields as required. By using md-is-error in the
input containers can immediately mark such fields as required (and
not after they have been touched, which is the default). To also make
the ng-message appear directly some extra styling was needed.

A failing save now always triggers a toast. When the save failed
because of invalid fields we show the ERROR_INVALID_DATA message.