CHANNELMGR-898 Fix red color of invalid multiple field
The use of md-is-error on md-input-container did not work for
multiple fields. When one of inputs in a multiple field became
invalid, all fields were colored red instead of only the invalid one.

We cannot use md-is-error here because the name of the form field to
check for validity requires the index of the field, which we don't
have since the ng-repeat is on the same md-input-container element.
Instead we override the styling of invalid fields and base it on
the .has-invalid-field class so fields marked as ng-invalid are
directly colored red.