HSTTWO-3879 Support for combined, first, second and stale cache stats over jmx
Note that also for the binaries cache and webfiles cache we now expose
cache stats. For now, the binaries cache and webfiles cache will never have a second
level cache or stale cache but this could be wired via Spring if we want to support that.

Next to the stats for first, second and stale cache, we also expose stats for the
misses and hits of the three combined. This is interesting to see because that is in the
end the effective cache hits/misses 'experienced' for the requests.