CHANNELMGR-1101 Hide iframe while loading a new scaled page
The contents inside the iframe is scaled because scaling the iframe
element itself results in ugly scrollbars. But the contents can only
be scaled once loaded. Showing unscaled contents first before scaling
it down results in visual flicker, so we hide the whole iframe and
show it again when the new contents in the iframe is ready.

Show the site with fadeIn() instead of show() to smoothen the
transition from seeing the canvas to seeing the site.