CHANNELMGR-1130 Fix scaling of dragged mirror
The mirror is already positioned exactly on top of the - possibly
scaled - original element. However, it is not scaled yet. So the
only thing that needs to be done is to scale the mirror down using
a 'transform-origin' of 'top left'. However, that will also scale
down the dimensions of the mirror, which are set explicitly by
Dragula. To fix this we first enlarge the mirror before scaling it
down again.

As a result, the scaling service only has to scale down the
iframe's HTML tag. The 'appElementsToScale' (which used to be only
the Dragula container) has been removed. Instead the drag-drop
service is responsible for scaling the mirror once it has been
created.

Setting the offset of the mirror wrapper seemed to be pointless
and has therefore been removed.