CMS-10585 Allow resource servlet to proxy from ROOT
If we set the publicPath of the frontend-build to "" - to make sure
all sources are required using relative paths - we need to be able
to configure the ResourceServlet proxy to serve from ROOT as well.

Example config:
- resource.proxies = angular/hippo-cm@to=http://localhost:9090/
- jarPath = /angular

Results in forwarding "/hippo-cm/somefile.js"
to "http://localhost:9090/somefile.js"

If a proxied resource is not found, we now prevent a NPE and
respond with a 404 instead.