CHANNELMGR-1118 Prevent drag wihout going through the overlay
In certain cases a bit of text can bleed out of the container overlay. This makes it possible (with precise mouse movement) to start dragging the related component while bypassing the overlay. This leads to a broken application state and is prevented by checking if we actually mousedowned on the overlay.