CHANNELMGR-1149 Only redraw component for changed props in active tab
Tabs that have just been deactivated sometimes still fire a
'propertiesChanged' event, probably because of a lingering
'change' event in the form. This results in a double re-rendering of
the component: first for the activated tab, and then immediately
also for the deactivated tab. As a result, the DOM changes of the
first re-rendering haven't been fully processed by the browser, and
the new markup (put by jQuery in a DOM fragment in a 'replaceWith'
call) is not part of the DOM tree yet. The second re-rendering then
tries to replace the first markup again, but because the parentNode of
the markup does not exist yet (i.e. is null because the DOM fragment
has not been attached to the tree yet) the second re-render call
fails.

Solution: avoid two quick consecutive renderComponent calls.

In this case the second re-rendering should not have taken place at
all (we only want to see the markup for the activated tab). The code
therefore now checks whether the form that generates the
'propertiesChanged' event is indeed the form of the active tab. Only
then the panel propagates the 'propertiesChanged' event.