CHANNELMGR-1149 Fix null parentNode in second renderComponent call
When renderComponent is called twice quickly, both calls have
fetched the component already and then fetch the component markup
in parallel. The first returned markup replaces the DOM of the
component. The second returned markup will also try to replace the
markup again, but since it still uses an old component reference
the markup of that old component is no longer attached to the DOM.
Hence the parentNode of the replaced elements is null, and the
replaceDOM call fails.

Fixed by ensuring that we replace the DOM of the latest component in
the page structure.