CHANNELMGR-1149 Always generate box element
Generate a box element when a component does not have any elements
(e.g. is rendered 'hidden'), also when a component is re-rendered.

We used to only generate box elements when an overlay element was
created. That failed when a component was re-rendered hidden in a
container of type 'HST.NoMarkup', because in these containers
components are assumed to contain a box element.

Added a unit test to cover this case.