CMS-9105 Don't encode URL value twice
When the URL value _is_ read from the URL component (because the
URL is editable) the URL has already been encoded. The new code
then encoded the URL again, which was not correct.

Also added a comment on why we read the URL value directly from
the name component's model when the URL is editable.