CHANNELMGR-1160 Fix save/restore scrollPosition functionality
Fixed the implementation of the save/restore scroll position functionality of the ScrollService and re-added the unit tests that went missing when we merged CM-Overlay and VE.

The initialisation of the jquery helper variables for the iframe has moved into a separate function. This function is called when the iframe loads, and used by the save/restoreScrollPosition functions, which can be called before the iframe onload handler is called (because a previous onload handler from a different component triggers a save/restore of the scroll position).

Lastly, the ugly code for retrieving the scrollTop position has been simplified by using iframeWindow.scrollTop() which is 'the most cross-browser way' to get the value: http://stackoverflow.com/questions/5371139/window-scrolltop-vs-document-scrolltop#comment20034307_5371386