CHANNELMGR-1115 Prevent visual editing when draft has invalid changes
Closing the full editor is now done before creating the draft instead
of afterwards. The close action checks whether the editor is valid
(i.e. the document does not contain any unsaved invalid changes).
The editor is only closed when the document is valid. The result is
reported back to the app. When the close failed (i.e. the editor is
not closed because of pending invalid changes) a message is shown
instead of the visual editor. Users can then click the 'switch to
full editor' link to open the full editor and correct the invalid
pending changes.