CHANNELMGR-1181 Make sure container min-height is present
Drag/drop code does not update the structureElement models before actually dropping, i.e. while drawing the shadow. This leads to containers without a min-height if they only contained a single item and said item was rendered in a different container (but not dropped yet).

Fixed it by checking the DOM instead of the structureElement model. Hst-container-items can be descendants of hst-containers, hence we have to use $.find() instead of $.children.