CHANNELMGR-1065 Use CKEditor URL and timestamp provided by the CMS
The CKEditor sources are loaded when the first HTML field is being
rendered. The CKEditorService takes care of loading and remembers
the promise so it's only loaded once.

The CKEditor URL and timestamp are retrieved from Wicket and
passed to the app. The URL is injected to the app using DomService
.addScript. It needs to be absolute, so it's prefixed with the CMS
context path (the Wicket URL is relative to the CMS app URL).

I've copied the loading mechanism of CKEditor used in the CMS.
Especially with the source version, loading may take a while and
the exponential backoff check is necessary.

The textarea is replaced by CKEditor, but we hide is with CSS
while the editor is loading to avoid showing raw HTML. Use
visibility instead of display to let the textarea take up the
right amount of space.