CHANNELMGR-1035 Disable custom config loading when not configured
CKEditor tries to load a 'config.js' file by default, which is not
present in our CKEditor fork, and leads to a 404. We therefore
disable the loading of such a 'custom config' file by setting
the config property 'customConfig' to an empty string, unless the
(customized) configuration indeed specifies a customConfig property.