CHANNELMGR-1227 Update angular-ui-tree
This version has fixes for horizontal drag movement, making it easier to mark a tree item as a child node, although it is now a bit too sensitive for my likes. But it is still better than how it previously handles horizontal movement which made it pretty hard to mark a tree item as a child.