HSTTWO-3936 Make sure for CMS requests also the thresholds work properly
Problem was that CMS requests results in a preview decorated VirtualHostsService which
is not an instance of a VirtualHostsService but of VirtualHosts. Hence, the instanceof would
fail for CMS requests, resulting in CMS requests not correctly taking the threshold (and some other) values into
account. Fixing it by moving some methods to the public interface (since #isDiagnosticsEnabled was already
public, I added the rest also to the public interface)