CHANNELMGR-1234 Include chosen rich text fields wrapped in compounds
Rich text fields can be chosen in 'choice' fields too. To code so
far assumed that all choices are compounds, but that's no longer
true: rich text fields are not compounds (also not according to
the content type service). They merely store their value in a
child node.

To distinguish 'compounds' and 'types that store their value in a
node instead of a property', the FieldType hierarchy got a new
member 'NodeFieldType' to express the latter. FieldType became an
interface, with AbstractFieldType as the base class. This allows
ChoiceFieldType, CompoundFieldType and RichTextFieldType to all be
a NodeFieldType while they inherit different classes. Also, the
code now expresses that choices have to be a NodeFieldType.

The frontend code expects that all choices are represented as
compounds. We therefore wrap a chosen rich text field in compound.
Initializing a compound used as a choice is now the job of the
CompoundFieldType, which got two new init() methods for this reason.

TODO:
 - add unit tests for HTML fields as compounds
 - include chosen rich text field in documents
   (right now they are only part of the document types)