HSTTWO-3959 Load channel nodes from below the hst:configuration node instead of below hst:channels
Note : Unit and integration tests still need to be fixed

With this refactoring, channel nodes are from now on stored below the hst configuration, for example instead of
+ hst:hst
  + hst:channels
  | + myproject
  | |  + hst:channelinfo
  + hst:configurations
    + myproject

as

+ hst:hst
  + hst:channels
  + hst:configurations
    + myproject
      + hst:channel
        + hst:channelinfo

note that the /hst:hst/hst:channels node for now I kept, because based on this node,
a cms user is webmaster or admin on channels.

Next this, I also added support for hst:channel inheritance: Aka, if the hst:configuration
is inherited from another channel, the channel info is also inherited. In this case
the channel settings in the channel manager are not editable.

Also, the hst:channel node has to be in the hst:workspace to be editable! Thus even if the
hst:channel is not inherited, it's settings are only editable if the channel info node is
in the hst:workspace, thus as follows:

+ hst:hst
  + hst:channels
  + hst:configurations
    + myproject
      + hst:workspace
        + hst:channel
          + hst:channelinfo

Since the hst:channel loading has now been completely changed, also caching is fixed: Aka,
channel loading now piggy backs on the same caching and invalidation mechanism as other
hst configuration objects.

Another improvement that has been done is the general hst preview creation: When creating
a preview, we only copy the hst:workspace from now on to preview, and have the preview
extend the live. Thus for example before we had:
+ hst:hst
  + hst:configurations
    + common
    + myproject (hst:inheritsfrom = ../common)
    |  + hst:sitemap
    |  + hst:pages
    |  + hst:components
    |  + hst:templates
    |  + hst:workspace
    + myproject-preview (hst:inheritsfrom = ../common)
       + hst:sitemap
       + hst:pages
       + hst:components
       + hst:templates
       + hst:workspace

From now on we have:

+ hst:hst
  + hst:configurations
    + common
    + myproject (hst:inheritsfrom = ../common)
    |  + hst:sitemap
    |  + hst:pages
    |  + hst:components
    |  + hst:templates
    |  + hst:workspace
    + myproject-preview (hst:inheritsfrom = ../myproject)
       + hst:workspace

Note the preview inheriting from ../myproject (thus live!)