HSTTWO-3982 Support for creating a branch of an hst configuration
when a branch is created, we set a mixin on the new hst:configuration node
that indicates that it is a branch. Also we add a String property containing the
name of the configuration it is a branch of. In the future, we can through this
also create branches of branches