CHANNELMGR-1261 Don't recreate whole editor markup after each save
All ng-repeats did not use 'track by', which made Angular use the
default behavior: a $$hashKey property per fieldValue object based on
an internal counter. Whenever the document was saved and the
server returned a new version, all the $$haskKey values would be
wiped and Angular would re-create all editor. This was noticeable
as a slow re-draw of the whole editor after each save. It also made
the scrollbar always jump to the top again.

We now track all field values by $index, and fieldTypes by id.
This ensures that all fields are rendered just once. An update of
the model (triggered by 'save') never re-renders any fields.

To ensure that CKEditor picks up the model change the ckeditor
component now uses $render to put the updated value into CKEditor.
This triggers a 'change' event, so to avoid a $digest loop we
don't update the viewValue again while we're processing the model
update.

TODO: add unit test for this to ckeditor.component.spec.js