CHANNELMGR-1262 Create preview config when channel is loaded
Creating the preview configuration moved from the channel
controller to the channel service. It used to be in the former
because a click on the 'Edit' mode button triggered it. But since
a channel is now - conceptually - always in 'edit' mode, it moved
to the latter.

When a new channel is loaded, HST preview configuration is created
immediately when:
  a) the user can write AND
  b) the channel does not have preview config yet

If it fails - e.g. when the hst:configuration -preview node is
missing but the hst:channel -preview node is still present - an
error message is shown and the original, non-editable channel
information is used. This results in a read-only version of the
channel manager, just like users without write access would have.

The error message ("Failed to enter Edit mode") is translated. But
while the first channel is being loaded, translations may not have
been fetched yet. To ensure that the translation are loaded in time,
the 'hippo-cm' root state is no longer abstract. That also caused
all unit tests to lookup translations, which failed when
ConfigService.locale was not defined. We therefore fall back
to the hardcoded 'fallback' language ('en') when calling
$translate.use().