CHANNELMGR-1262 Remove custom parent container of toast messages
We need a toast message when creating preview config fails. In
that case hippo-iframe (the default container) is not rendered yet
when the channel is opened for the first time. The mdToast service
then falls back to using the body, but the second time (when
hippo-iframe is rendered) it does not show toasts anymore.
Probably a bug.

But: all toast messages should be shown just below the main
toolbar. So instead of having complex logic around the parent
element, move the toast container a little down with CSS.