CHANNELMGR-1145 Add 'Not all fields are shown' message
Since 'switch to the full editor' is a clickable link the
translations need to be split into three parts. The last part (at
the end of the sentence) is empty in English, but may be needed in
other languages.

Instead of adding a 'Got it' button to hide the message, the
message is now part of the scrollable content. Scrolling down in
the editor will thus hide the message automatically, which seems
more natural than clicking the button each time.