CHANNELMGR-1299 Fix validation check for input fields
The 'name' attribute of input fields was not interpolated and
hence not set correctly, which broke the lookup of the error for a
field.

Also fixed the interpolation of the maxlength and aria-label
attributes of input fields, which were not interpolated either.
Note that the aria-label fields of textarea and ckeditor *are*
interpolated so these are fine.

The styling for the .has-focused-field was more specific than
that for the .has-invalid-field, which caused an invalid input field
that has focus to keep a blue instead of a red label when the field
became invalid.

Moved the 'isValid' check from the documentFields controller to
the primitiveField controller since it was only used in the latter.