HHP-4 Fix unwanted deletion+recreation of facets
The FacetVisitor uses the new before/after hooks to map the
current facets on a node by facetName -> targetUUID. This map is used
to lookup a facet name by id, or viceversa, and to mark facets that
have been visited, i.e. are present in the DOM tree.

When the visitor has finished visiting the DOM, the map is check for
unmarked facets, which are then deleted.