REPO-1685 Add and fix implementation of (WEAK)REFERENCE value mapping from/to relative or absolute paths
Using absolute or relative paths is only (and only needed to be) based on the current ContentDefinition#rootPath.
If the referred path is not within that root path and abolute path will be used, otherwise the relative path from the root path.
Mapping a relative path to the target (path) reference therefore only is dependent on the content definition root path.

Which doesn't mean it actually can resolve the path (it might not yet be bootstrapped for example), in which case the reference will be dropped.
But that problem is independent of using a relative or absolute path, and simply cannot be 'fixed' or prevented.
That only depends on how and which content is (auto|manual) exported and in which 'chunks'.