ARCHE-539 Mark certain parts of the hst configuration to have 'content' residual children
Note that on purpose, I did not change sitemenus.yaml and container.yaml since the
residual 'content' children should be on a lower level.

Also it is debatable whether the 'hst:channelinfo' in channel.yaml is actually content
or configuration. I think for practicality reasons we can best keep it configuration, just
like its parent node.