CHANNELMGR-1380 Fix locking md-select by calling $mdSelect.hide()
Its a workaround for an angular-material issue:
https://github.com/angular/material/issues/10747

Instead of using the $timeout solution initially proposed I decided to
use the (undocumented) $mdSelect.hide() promise as described in the
github issue comments. It's a cleaner solution because it does not
depend on a parameter (100 ms.) and it is applied close to the root
of the origin, namely the projectToggle instead of the channel service.