CHANNELMGR-1368 Confirm discard changes if channel closed via breadcrumb
The 'Channels' entry in the breadcrumb no longer switches to the
Channel Overview card directly. Instead, it sends a 'close-channel'
event to the app. The app can then close the right side-panel, and
confirm the discarding of changes when needed. Once OK, the app sends a
'close-channel' event back to the CMS, which will show the Channels
Overview.

The channel actions menu now listens to the 'close-channel' event from
the CMS, so closing the Channel via the breadcrumb triggers the same
code as selecting 'Channel > Close' in the menu.

The breadcrumb toolbar now checks whether the click handler of the
toolbar item returns false. If so, the breadcrumb is not updated. This
is now used by the first breadcrumb item, which waits for the response
from the app to update the breadcrumb.