HSTTWO-4121 The mount property hst:nochannelinfo can better not be inherited by child mounts
currently, when you have

+ hst:hst
   + hst:hosts
        + dev-local
            + localhost
                + hst:root
                   - hst:nochannelinfo = true
                   + en   [hst:mount]
                   + nl    [hst:mount]

Then because the hst:root has

hst:nochannelinfo = true

also the mounts 'en' and 'nl' get this property inherited currently.
This can better be not inherited because in some projects, the hst:root
is a 'placeholder' mount that does not have a channel. That would mean
that all child mounts would require explicit

hst:nochannelinfo = false

That is unhandy. Better to not inherit the property