HIPPLUG-1486 Move config parameters of imagelink to _default_ node
Move all default values of the image picker parameters to the _default_
node of the imagelink type. Add them to frontend:properties so end users
can adjust the default values in the document type editor.

The preview.width property has been changed from 50 to 60 because the
thumbnails are 60x60 by default. Rendering them 50px wide make the
thumbnail images unnecessary blurry.

The validator.id property could not be set on the _default_ cluster
as-is because this property is treated special by the plugin framework.
The default value property has been called 'image.validator.id' instead.