HIPPLUG-1486 Make ImageItem independent of Wicket
Introduce a JcrSessionProvider that can be used by the ImageItem to
obtain a JCR session instead of hardcoding getting it from Wicket's user
session. Add PowerMock to be able to intercept the call to
JcrHelper.getPrimaryItem().

Removed the dependency on commons-lang since it's not used anywhere.

Also enable the happy path unit test again. Why was that disabled?