REPO-1811 Support for a Database lock manager
The Database lock manager interaction is pretty much exactly the same
as the MemoryLockManager interaction. The biggest difference is that
the DbLockManager creates database based locks and the MemoryLockManager
creates MemoryLocks. Note that a DbLock is the same as the MemoryLock
but only contains a #destroy implementation that releases the database
lock.

The DbLockManager is still work in progress