CMS-10910 Add methods for CmsSessionContext attributes management
Add get/set/remove methods for the management of session like
attributes. These attributes are meant for external use, e.g. in case
a project needs to manage cross-webapp user settings.
Since the CmsSessionContext is shared amongst web-apps the values of
these attributes must be either be JVM classes or classes that are on
the shared classpath.