CMS-10913 Add generic dialog manager
- manager provides an AjaxBehavior to invoke the dialog from the client.
- uses a configProvider to provide a stateful and stateless
configuration of the dialog.
- provides a hook #onConfigure to be able to extract data from the
request that invokes the dialog

This can all be simplified if we refactor the existing dialog behaviors,
like the ImagePickerBehavior (used by ckeditor) into managers. It means
we move most of the logic and state in DialogBehavior (getting the
DialogService, setting close and cancel actions, parsing dialog
configuration) into the DialogManager and simplify the DialogBehavior
to invoke a callback with parsed request parameters to the manager when
the dialog needs to open.