CMS-10897 introduce LockManagerException for some methods
For a method like #isLocked or #getLocks, in case of an exception, we
need to throw an exception: Otherwise we need to return true|false or
an empty list: Client code can't handle this right. Then we can choose
between a runtime or checked exception. Since something like #getLocks
is something a client can react on (for example show a failed message in
UI for lock overview or a retry), a checked exception makes most sense