REPO-1835 Refactor DocumentWorkflowImpl
Goal is to be able to set an initializationPayload via the
DocumentWorkflow interface to all actions and hints.

Introduce DocumentWorkflowTransition with builder to facilitate
various combination of parameters that have to be passed on the
triggerAction method of the ScxmlExecutor.

Do not refactor the request... methods for now. Refactoring would need
an interface for Request or more state in the DocumentWorkflowTransition
and subsequent logic in the transition method.

Add Enum for action to avoid typos in calls to api.

The initializationPayload always contains the key
org.onehippo.repository.documentworkflow
.DocumentWorkflowTransition.Builder.INITIALIZATION_PAYLOAD_USED

If the setter for initializationPayload has not been called the value
is false otherwise true. In this way the workflow is able to determine
if the initializationPayload is set by the caller of the api.