REPO-1835 Add requestIdentifier to builder
Add requestIdentifier to builder in order to set the
event payload key, value pair on the eventpayload and the allowed
actions a started scxmlExecutor is necessary. If it has not been
started, the feedback containing the requests has not been initialized
yet.

An additional benefit of adding the requestIdentifier to the builder
is that callers of the api can also use workflowmethods that request
a requestModifier using the new transitionMethod in combination
with the DocumentWorkflowTransition.Builder.

This approach seems preferably over introducing Futures, since the
current cases of using a request Identifier all result in the same
modification of the actionsMap and the eventPayload. Using Futures
furthermore makes the api overly complex, without adding any new
functionality.

It's still possible to set the actionsMap as well. The available actions
derived from the requestIdentifier are added in that case.