REPO-1835 Detect if initialPayload has been set
Do not initialize initialPayload so that the scxml groovy script
can detect if it has not been set. If it has not been -> there are
no actions on the document ( fail fast ). All plugins should
supply an initial payload to the workflow if they want to use the
WPM feature.