REPO-1859 Add new default implementation to Workflow
Add transition method to Workflow interface, because both
DocumentWorkflow and EditableWorkflow use that method.

Add default implementation to assure backward compatibility.

Add interface and unit test with testcase, because:
1. Custom actions can use their own Enum;
2. DocumentWorkflowAction does not need to be exposed by the API

Move constants from DocumentWorkflow to DocumentWorkflowConstants as
part of the API.

Removed hints(payload) from GalleryWorkflowImpl, because we now have
a default implementation.

Renamed DocumentWorkflowTransition to WorkflowTransition, because it
does not only apply to the DocumentWorkflow.