REPO-1835 Do not override default implementation in a sub interface (ever)
The DocumentWorkflow extends Workflow. That DocumentWorkflow redefines
#hints() with other javadoc is imho already bad practice but has been like
this for a long time

Now that we've added #hints(initialPayload) to Workflow with a default
implementation, we should never redefine this method in a sub interface
with a different javadoc: I understand we might want different javadoc,
but overriding in an interface a default implementation can result in
havoc: If your implementation implements both interfaces, it becomes
really spooky which default method is used. Most probably the subinterface
hides the super interface hints completely, but it is quite delicate and
there might be subtle issues. When a sub interface redefines the same
method but now without default impl, I really don't know what the end
result is, but frankly, I don't want to know because it is a pattern we
should always avoid