CMS-10951 Use StringCodecService instead of the 'codecs' config node
The configuration of (possibly locale-specific) StringCodecs for
encoding display names and node names has moved to the
StringCodecService in the repository. The new configuration location is
/hippo:configuration/hippo:modules/stringcodec/hippo:moduleconfig.

Because the ISettingsService is public API, we still need to be able to
provide a StringCodecFactory. Hence the ServicedStringCodecFactory, who
does not lookup codecs in its own map but delegates lookups to the
StringCodecService instead.