REPO-1870 Avoid long stacktraces being logged by default
When hitting a non existing endpoint, or an endpoint with the wrong
method, by default, a long stacktrace is logged.

By default, we should log a single line warning and in debug level the
whole stacktrace.

Unfortunately using:

 final WebApplicationExceptionMapper webApplicationExceptionMapper = new WebApplicationExceptionMapper();
 webApplicationExceptionMapper.setPrintStackTrace(false);
 endpointFactory.setProvider(webApplicationExceptionMapper);

doesn't do the job because then nothing is logged any more (unless in FINE
log level).

Hence extend the WebApplicationExceptionMapper, log a warning and
delegate