HSTTWO-4152 Know whether componentParameter of manageContent is relative
When the manageContent tag has a componentParameter attribute, the tag
figures out whether that parameter is supposed to store a relative path
or an absolute path. It looks up the annotated method for the parameter
in the parameters info class of the current component. When that method
has a JcrPath annotation with isRelative=true the component parameter is
supposed to be relative, otherwise it's supposed to be an absolute path.

To be able to get the current component in the tag, we set the
HstComponentWindow as a request attribute (similar to how HstRequest and
HstResponse are exposed).