CHANNELMGR-1553 Return slugs as media type "text/plain"
The returned string is not valid JSON because it's not quoted. Yet
frontend libraries will try to parse the result as JSON, which fails.

Returning a slug as plain text makes it impossible to return an
ErrorInfo object, but since creating a slug never generates an error
anyway that's not a problem.