CHANNELMGR-1481 Add "folders/{path}" endpoint
Retrieves a list of Folder instances based on the provided path. A
Folder instance contains a name, displayName, locale and path. All
folders in the provided path are considered valid. If a node referenced
by the path exists, the displayName and locale are copied from it. If
the locale is null, it will be copied from the parent.
In case of a non-existing node, the displayName will equal the input
value and the locale will be copied from the parent, or null.