CHANNELMGR-1543 Fix closing of edit-content panel
The closePanel() method removed in the previous commit was used by the
edit-content panel, so it had to be added back. Added a on-close
callback to the edit-content component so it can trigger the
closePanel() method in the same way as the step-1 component instead of
calling the SidePanelService directly.

Still TODO: adjusts the unit tests.