CHANNELMGR-1511 Also delete newly created drafts
DocumentWorkflow.delete() actually archives a document, e.g. moves it
to the attic, stripped of all data, to keep a pointer into version
history. That's fine for existing documents that are allowed to be
deleted (e.g. an offline document).

But: DocumentWorkflow.delete() is also not allowed for newly created
drafts. To delete those, FolderWorkflow.delete() can be used. Which has
the nice bonus that it does not archive the deleted document, which is
not needed for new drafts anyway. We therefore use the term "erase" to
make a distinction from "delete".

However, FolderWorkflow.delete() also deletes any existing document
without problems. That's a bit dangerous, so we only want to use it for
newly created drafts. We therefore check whether the document has a
preview variant. If so, we don't allow the document to be deleted via
the FolderWorkflow (it should have been deleted by the DocumentWorkflow
already). Only documents without a preview variant are then
actually erased.