CMS-10946 Keep "hippo:username" session attr until last app logged out
The HTTP session attribute "hippo:username" is used by the
WhitelistedClassesResourceGuard to check whether a user is logged in or
not. When logged in, all resources can be accessed, else only resources
in whitelisted packages.

The logout method used to always remove the "hippo:username" attribute
from the HTTP session. That broke behavior when two applications share
the same HTTP session, as is the case with CMS and Console. When a user
was logged in to both CMS and Console and then logged out from one of
them, the "hippo:username" attribute got removed from the HTTP session
and the other app could no longer access non-whitelisted resources.