CHANNELMGR-1568 Set display name and url name when updating draft
Both still need to be encoded by the display name and node name encoder,
respectively. That's also relevant when creating a new document, because
the names need to be encoded before checking their existence.

The encoding logic for display names and url names is shared, so renamed
DisplayNameUtils to DocumentNameUtils and put all logic in there. The
#setNames method can be used to set both simultaneously, which prevents
fetching the workflow twice.

Removed #getDisplayName, since the repository equivalent
DocumentUtils.getDisplayName() can be used instead.