CHANNELMGR-1568 Update display name and URL name in separate PUT call
Allows updating the names of a document without side-effects. Updating
the names as part of the draft call would also save all fields and
create a preview version of the document.

Updating document names accepts a document. Only the displayName and
urlName properties in it are actually used. Both are encoded again by
the backend, so the frontend can pass the not-yet-encoded version of
these fields (mainly relevant for manually set URL names).

Use the DocumentWorkflow to set the URL name, so the workflow prevents
illegal cases (e.g. renaming a published document). The exception is
renaming an new draft document (e.g. without a preview version yet),
which is not possible via DocumentWorkflow. For that specific case the
DefaultWorkflow is used, which always allows renaming.

The utility methods to retrieve a certain workflow have been moved to a
separate class ContentWorkflowUtils, which is now used by both the
DocumentServiceImpl and the DocumentNameUtils.