CHANNELMGR-1568 Don't update the URL name if the display name clashes
First check that neither the URL name nor the display name clash
with sibling documents before updating either of them. Otherwise the
case where both are changed but the display name clashes is not handled
correctly. The URL may already have been updated via workflow before an
error is returned because the display name clashes with an existing
node. Because the URL is updated via workflow in a separate session,
the new URL name would already have been persisted. That should not
happen: either both the display name and url name are changed, or
neither of them are.

Added a unit test for this case, and reshuffled the existing tests a bit
to better reflect the new code ordering.