CHANNELMGR-1568 Accept only ErrorInfo or Document as error payload
A string payload won't be parsed as valid JSON by the front-end.

Changed the signature of all concrete ErrorWithPayloadExceptions to only
accept an ErrorInfo object, so invalid cases (like a string) are
caught by the compiler. The exception was the BadRequestException, which
is also used by DocumentServiceImpl#updateDraft to return a document as
payload. Added a second constructor for that case.

This exercise also found two erroneous cases of an
InternalServerErrorException that got an exception as payload
(in DocumentServiceImpl#archiveDocument and #eraseDocument). Such cases
will now be avoided.