ESSENTIALS-1113 Remove unused undo and display location concepts
The Essentials API's PluginEvent used to expose a concept of 'undo'
and 'display location'. Both concepts have never been in use in the
productized version of Essentials. The concept of undoing plugin
installation has been discussed in the past, but was agreed to be
too difficult to implement. The concept of a display location seems
to refer to initial UI design ideas, which have never been implemented.
Since both concepts are not in use, they constitute dead code, which
is why I'm removing it.