ESSENTIALS-1113 Remove support for DisplayEvent
DisplayEvent seems an outdated / incomplete approach to send messages
including HTML formatting information to the front-end. the events are
posted on the eventBus, and MemoryPluginEventListener would collect
them in a list, which was only ever accessed by unit test code.

DisplayEvent's DisplayType enum was also referenced (and stored) in
MessageRestful classes, and as such, they could have been serialized
and sent to the front-end, but the front-end code makes no use of that
additional information. In the absence of any kind of documentation of
the intent of this mechanism, I can only conclude that this is an
outdated, dead mechanism, which is why I'm cleaning it up.

The clean-up work on PluginEvents so far has triggered a move of some
message templates from a properties file (from which they were assigned
to String instance variables through Spring) into the log messages
directly, and so I've also removed the obsolete Spring configuration.