ESSENTIALS-1113 Uniform usage of PluginContextFactory
As a preparation to turn the PluginContextFactory into an injectable
Spring bean, I make sure that all usages of PluginContextFactory
outside of the dashboard module are uniform: For plugins, the factory
is used immediately when a REST resource is invoked. For SDK utilities,
the factory is not used, but the PluginContext is passed in as an
argument.

Also, I replace some trivial helper classes with their less boiler-
plate intensive Java 8 lambda counterparts.