ESSENTIALS-1113 PluginContextFactory as Spring Bean
The PluginContextFactory is now a injected as a Spring Bean into the
classes (mostly REST resources) which need to create a new PluginContext.
The dashboard still has a few calls to PluginContextFactory#getContext
done in (deeply) nested method call stacks, which probably constitutes
a piece of inefficiency of Essentials.
With this change in place, the only thing that keeps us from moving the
DefaultPluginContext out of the SDK implementation module and into the
dashboard module is the fact that the SDK implementation module extends
it for its unit tests.