ESSENTIALS-1113 Move PluginContextFactoryImpl back to SDK
I realized that my goal should not be to get rid of the
plugin-sdk/implementation, but to get rid of the need for Essentials
plugins to depend on it, rather than on the plugin-sdk/api only. API
and implementation should have the same lifecycle, which should be
distict from the lifecycle of the dashboard, such that we can eventually
move the plugins to their appropriate location (outside of the
Essentials project).